requirements.txt
yt-dlp>=2024.1.0
openai>=1.0.0
elevenlabs>=1.0.0
pydub>=0.25.1
python-dotenv>=1.0.0
rich>=13.0.0
